Backend_vulnerable/requirements.txt
fastapi==0.100.0           # Framework for building APIs
uvicorn==0.23.0            # ASGI server for running FastAPI
mysql-connector-python==8.0.33  # Connector for MySQL database
loguru==0.7.0              # Logging library
python-dotenv==1.0.0       # Manage environment variables
pydantic
python-decouple
python-multipart